bug fix exact match filter and add/refcator policies
