+++ oligo_designer_toolsuite/oligo_specificity_filter/_filter_base.py
@@ -124,7 +124,7 @@ class AlignmentSpecificityFilter(SpecificityFilterBase):
                 region_id=region_id,
                 oligo_database=oligo_database,
                 file_index=filename_reference_index,
-                consider_hits_from_input_region=True,
+                consider_hits_from_input_region=False,  # for filter we do not want to remove hits within the same region
             )
             for region_id in region_ids
         )
@@ -178,7 +178,7 @@ class AlignmentSpecificityFilter(SpecificityFilterBase):
                 region_id=region_id,
                 oligo_database=oligo_database,
                 file_index=filename_reference_index,
-                consider_hits_from_input_region=True,
+                consider_hits_from_input_region=True,  # for cross-hybridization we also want to remove hits within the same region
             )
             for region_id in region_ids
         )
@@ -189,7 +189,6 @@ class AlignmentSpecificityFilter(SpecificityFilterBase):
 
         table_hits = pd.concat(table_hits, ignore_index=True)
         oligo_pair_hits = list(zip(table_hits["query"].values, table_hits["reference"].values))
-        print(oligo_pair_hits)
 
         return oligo_pair_hits
