+++ oligo_designer_toolsuite/oligo_specificity_filter/_filter_bowtie.py
@@ -183,7 +183,7 @@ class BowtieFilter(AlignmentSpecificityFilter):
         :return: A tuple containing a DataFrame of significant Bowtie hits and a list of oligos with these hits.
         :rtype: (pd.DataFrame, list)
         """
-        if consider_hits_from_input_region:
+        if not consider_hits_from_input_region:
             # remove all hits where query and reference come from the same region
             search_results = search_results[
                 search_results["query_region_id"] != search_results["reference_region_id"]
@@ -368,7 +368,7 @@ class Bowtie2Filter(AlignmentSpecificityFilter):
         :return: A tuple containing a DataFrame of significant Bowtie2 hits and a list of oligos with these hits.
         :rtype: (pd.DataFrame, list)
         """
-        if consider_hits_from_input_region:
+        if not consider_hits_from_input_region:
             # remove all hits where query and reference come from the same region
             search_results = search_results[
                 search_results["query_region_id"] != search_results["reference_region_id"]
