+++ oligo_designer_toolsuite/oligo_specificity_filter/_filter_exact_matches.py
@@ -2,15 +2,18 @@
 # imports
 ############################################
 
-from typing import List, get_args
-
 import iteration_utilities
 import pandas as pd
+import networkx as nx
+
+from typing import List, get_args
 from joblib import Parallel, delayed
 
 from .._constants import _TYPES_SEQ
 from ..database import OligoDatabase, ReferenceDatabase
-from . import SpecificityFilterBase
+
+from ._filter_base import SpecificityFilterBase
+from ._policies import RemoveByLargerRegionPolicy
 
 ############################################
 # Oligo Exact Match Filter Classes
@@ -45,35 +48,42 @@ class ExactMatchFilter(SpecificityFilterBase):
         ), f"Sequence type not supported! '{sequence_type}' is not in {options}."
 
         # extract all the sequences
-        search_results = self._get_duplicated_sequences(sequence_type, oligo_database)
+        sequences = oligo_database.get_sequence_list(sequence_type=sequence_type)
+        search_results = self._get_duplicated_sequences(sequences)
 
         # get mapping from sequence to oligo_id
+        # use same sequence as sequence_type
         sequence_oligoids_mapping = oligo_database.get_sequence_oligoid_mapping(
             sequence_type=sequence_type, sequence_to_upper=True
         )
 
         region_ids = list(oligo_database.database.keys())
-        results = Parallel(n_jobs=n_jobs)(
+        table_hits = Parallel(n_jobs=n_jobs)(
             delayed(self._run_filter)(
                 sequence_type=sequence_type,
                 region_id=region_id,
                 oligo_database=oligo_database,
                 search_results=search_results,
                 sequence_oligoids_mapping=sequence_oligoids_mapping,
+                consider_hits_from_input_region=False,
             )
             for region_id in region_ids
         )
 
-        # Process results
-        table_hits, oligos_with_hits = zip(*results)
+        table_hits = pd.concat(table_hits, ignore_index=True)
+        oligo_pair_hits = list(zip(table_hits["query"].values, table_hits["reference"].values))
 
-        for region_id, oligos_with_hits_region in zip(region_ids, oligos_with_hits):
+        policy = RemoveByLargerRegionPolicy()
+        oligos_with_hits = policy.apply(
+            graph=nx.from_edgelist(oligo_pair_hits), oligo_database=oligo_database
+        )
+
+        for region in region_ids:
             database_region_filtered = self._filter_hits_from_database(
-                database_region=oligo_database.database[region_id],
-                oligos_with_hits=oligos_with_hits_region,
+                database_region=oligo_database.database[region],
+                oligos_with_hits=oligos_with_hits[region],
             )
-            oligo_database.database[region_id] = database_region_filtered
-
+            oligo_database.database[region] = database_region_filtered
         return oligo_database
 
     def get_oligo_pair_hits(
@@ -101,35 +111,38 @@ class ExactMatchFilter(SpecificityFilterBase):
         assert (
             sequence_type in options
         ), f"Sequence type not supported! '{sequence_type}' is not in {options}."
+        sequence_type_reverse_complement = options[0] if options[0] != sequence_type else options[1]
 
         # extract all the sequences
-        search_results = self._get_duplicated_sequences(sequence_type, oligo_database)
+        sequences = oligo_database.get_sequence_list(sequence_type=sequence_type)
+        sequences_rc = oligo_database.get_sequence_list(sequence_type=sequence_type_reverse_complement)
+        search_results = self._get_duplicated_sequences(sequences + sequences_rc)
 
         # get mapping from sequence to oligo_id
         sequence_oligoids_mapping = oligo_database.get_sequence_oligoid_mapping(
-            sequence_type=sequence_type, sequence_to_upper=True
+            sequence_type=sequence_type_reverse_complement, sequence_to_upper=True
         )
 
         region_ids = list(oligo_database.database.keys())
-        results = Parallel(n_jobs=n_jobs)(
+        table_hits = Parallel(n_jobs=n_jobs)(
             delayed(self._run_filter)(
                 sequence_type=sequence_type,
                 region_id=region_id,
                 oligo_database=oligo_database,
                 search_results=search_results,
                 sequence_oligoids_mapping=sequence_oligoids_mapping,
+                consider_hits_from_input_region=True,
             )
             for region_id in region_ids
         )
 
         # Process results
-        table_hits, oligos_with_hits = zip(*results)
         table_hits = pd.concat(table_hits, ignore_index=True)
-        hits = list(zip(table_hits["query"].values, table_hits["reference"].values))
+        oligo_pair_hits = list(zip(table_hits["query"].values, table_hits["reference"].values))
 
-        return hits
+        return oligo_pair_hits
 
-    def _get_duplicated_sequences(self, sequence_type: _TYPES_SEQ, oligo_database: OligoDatabase):
+    def _get_duplicated_sequences(self, sequences: list):
         """Identifies duplicated sequences within an oligonucleotide database for a specified sequence type.
 
         :param sequence_type: The type of sequences to analyze within the database.
@@ -139,8 +152,7 @@ class ExactMatchFilter(SpecificityFilterBase):
         :return: A list of duplicated sequences found in the database.
         :rtype: list
         """
-        # extract all the sequences
-        sequences = oligo_database.get_sequence_list(sequence_type=sequence_type)
+        # convert to upper sequence
         sequences = [sequence.upper() for sequence in sequences]
 
         # find the duplicates within the database
@@ -157,6 +169,7 @@ class ExactMatchFilter(SpecificityFilterBase):
         oligo_database: OligoDatabase,
         search_results: List,
         sequence_oligoids_mapping: dict,
+        consider_hits_from_input_region: bool,
     ):
         """Filters oligonucleotides based on sequence duplications, identifying hits within and across regions.
 
@@ -170,23 +183,32 @@ class ExactMatchFilter(SpecificityFilterBase):
         :type search_results: List
         :param sequence_oligoids_mapping: Mapping from sequences to oligo IDs.
         :type sequence_oligoids_mapping: dict
+        :param consider_hits_from_input_region: Flag to indicate whether hits from the input region should be considered.
+        :type consider_hits_from_input_region: bool
         :return: A tuple containing a table of hits and a list of oligos with those hits.
         :rtype: (pd.DataFrame, list)
         """
 
         database_region = oligo_database.database[region_id]
-
         hit_dict = {}
         for oligo_id in database_region.keys():
             oligo_seq = database_region[oligo_id][sequence_type].upper()
             if oligo_seq in search_results:
-                # find all oligos with the same sequence
-                hits = [
-                    oligo_id
-                    for oligo_id in sequence_oligoids_mapping[oligo_seq]
-                    if not oligo_id.startswith(region_id)
-                ]
-                hit_dict[oligo_id] = hits
+                # find all reverse complements with the same sequence
+                if oligo_seq in sequence_oligoids_mapping:
+                    if consider_hits_from_input_region:
+                        hits = [oligo_id for oligo_id in sequence_oligoids_mapping[oligo_seq]]
+                    else:
+                        hits = [
+                            oligo_id
+                            for oligo_id in sequence_oligoids_mapping[oligo_seq]
+                            if not oligo_id.startswith(region_id)
+                        ]
+                else:
+                    hits = []
+                # don't store when all hits come from the same region and are filtered out
+                if hits:
+                    hit_dict[oligo_id] = hits
 
         table_hits = pd.DataFrame(
             [(key, value) for key, value in hit_dict.items()],
@@ -194,6 +216,4 @@ class ExactMatchFilter(SpecificityFilterBase):
         )
         table_hits = table_hits.explode("reference", ignore_index=True)
 
-        oligos_with_hits = list(set(hit_dict.keys()))
-
-        return table_hits, oligos_with_hits
+        return table_hits
